pi/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
pydantic-settings==2.4.0
aiofiles==24.1.0
aiosqlite==0.20.0
pycups==2.0.1
